refactor(drive-abci): move contested-resource decode helper into its own module
Move `decode_serialized_index_value` and `MAX_ENCODED_INDEX_VALUE_BYTES` (plus
their unit tests) out of `query/voting/mod.rs` into a dedicated
`query/voting/decode_index_value.rs` module, keeping `mod.rs` limited to module
declarations. The helper is re-exported from `voting` so call sites are
unchanged. No behavior change.
